Seperated handler from router
