+++ src/server/routers/questionnaireRouter.ts
@@ -1,5 +1,6 @@
-import BlaiseApiClient, { Questionnaire } from "blaise-api-node-client";
-import express, { Request, Response, Router } from "express";
+import BlaiseApiClient from "blaise-api-node-client";
+import express, { Router } from "express";
+import { QuestionnaireHandler } from "../handlers/QuestionnaireHandler";
 
 export default function questionnaireRouter(blaiseApiClient: BlaiseApiClient): Router {
     const router = express.Router();
@@ -7,20 +8,4 @@ export default function questionnaireRouter(blaiseApiClient: BlaiseApiClient): R
 
     router.get("/api/questionnaires", questionnaireHandler.getQuestionnaires);
     return router;
-}
-
-export class QuestionnaireHandler {
-    blaiseApiClient: BlaiseApiClient;
-
-    constructor(blaiseApiClient: BlaiseApiClient) {
-
-        this.blaiseApiClient = blaiseApiClient;
-        
-        this.getQuestionnaires = this.getQuestionnaires.bind(this);
-    }
-
-    async getQuestionnaires(_request: Request, response: Response<Questionnaire[]>): Promise<Response> {
-        const questionnaires = await this.blaiseApiClient.getQuestionnaires('gusty');
-    return response.status(200).json(questionnaires);
-    }
 }
